# Conflicts:
#	public/plugins/bootstrap-wysihtml5/bootstrap3-wysihtml5.all.js
#	public/plugins/bootstrap-wysihtml5/bootstrap3-wysihtml5.all.min.js
#	public/plugins/bootstrap-wysihtml5/bootstrap3-wysihtml5.css
#	resources/views/interventi/partials/consulente.blade.php
#	resources/views/interventi/partials/intervento.blade.php
#	resources/views/interventi/partials/progettocliente.blade.php

+++ app/Http/Controllers/AttivitaController.php
@@ -4,6 +4,8 @@ use App\Http\Requests\AttivitaRequest;
 use App\Attivita;
 use App\Progetto;
 
+use Illuminate\Support\Facades\Auth;
+
 use Illuminate\Support\Facades\Response;
 use Illuminate\Support\Facades\Input;
 class AttivitaController extends Controller {
@@ -34,6 +36,9 @@ class AttivitaController extends Controller {
      */
     public function store(AttivitaRequest $request)
     {
+        if(!(Auth::User()->consulente->tipo == 'Partner' OR Auth::User()->consulente->tipo == 'Admin')){
+            abort(503, 'Unauthorized action.');
+        }
         $data = $request->all();
         Attivita::create($data);
         if ($request->parent_id != 0) $attivita = Attivita::findOrFail($request->parent_id);
@@ -72,7 +77,10 @@ class AttivitaController extends Controller {
      */
     public function update(AttivitaRequest $request)
     {
-        $attivita = Attivita::findOrFail($request->parent_id);
+        if(!(Auth::User()->consulente->tipo == 'Partner' OR Auth::User()->consulente->tipo == 'Admin')){
+            abort(503, 'Unauthorized action.');
+        }
+        $attivita = Attivita::findOrFail($request->selected);
         $progetto = Progetto::findOrFail($request->progetto_id);
         $attivita->descrizione = $request->descrizione;
         $attivita->save();
@@ -88,6 +96,9 @@ class AttivitaController extends Controller {
      */
     public function destroy($id)
     {
+        if(!(Auth::User()->consulente->tipo == 'Partner' OR Auth::User()->consulente->tipo == 'Admin')){
+            abort(503, 'Unauthorized action.');
+        }
         $attivita = Attivita::findOrFail($id);
         $progetto = Progetto::findOrFail($attivita->progetto_id);
         $attivita->delete();
